Fair random testing
Quite a lot of redesign so that we can do fair tests. I was finding that my accuracy decreased as my training data increased. I had a suspicion that this was because I was obtaining the early parts of the training data from the same source as the testing data (there was no overlap between the actual files, but nonetheless the early parts of the training data might have come from the same source, and even the same speaker). So I've done quite a bit of moving things around so that the training data is taken randomly (but consistently, so that a larger dataset will contain all the files from a smaller one) from all the GALE recordings. We will see if that helps.