I *think* I'm now doing big experiments correctly
Fixed various things to make sure that when we run a big experiment we use the same testset, with the same grammar, and that everything is properly chosen from the right places. Still get the weird effect that increasing the training data makes things worse, but at least I am fairly sure that this isn't happening because I'm doing anything obviously stupid.