Trying to fix it so that we do SAMA properly, without breaking the others
This has been a nightmare: o do SAMA properly, we need madafiles.json to be in "old" Buckwalter. But then MADAMIRA and LDC need new Buckwalter. And then LDC has a weird grapheme "U" which breaks things, so I've simply substituted "u" for "U". And it's all been a horrible mess.